# projects/bot_eval_run/debug_node_raw_off_batch3/node-api-oracle_run_03/workspace/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TODO API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Check server health",
        "responses": {
          "200": {
            "description": "Server is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "summary": "Get OpenAPI specification",
        "responses": {
          "200": {
            "description": "OpenAPI specification",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/todos": {
      "get": {
        "summary": "Get all TODO items",
        "responses": {
          "200": {
            "description": "List of TODO items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "todos": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Todo"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new TODO item",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "TODO item created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          }
        }
      }
    },
    "/todos/{id}": {
      "get": {
        "summary": "Get a TODO item by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "TODO item found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            }
          },
          "404": {
            "description": "TODO item not found"
          }
        }
      },
      "patch": {
        "summary": "Update a TODO item by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "TODO item updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            }
          },
          "404": {
            "description": "TODO item not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a TODO item by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "TODO item deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "404": {
            "description": "TODO item not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Todo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "done": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "doneAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TodoRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          }
        },
        "required": ["title"]
      },
      "TodoResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "todo": {
            "$ref": "#/components/schemas/Todo"
          }
        }
      },
      "TodoUpdateRequest": {
        "type": "object",
        "properties": {
          "done": {
            "type": "boolean"
          }
        },
        "required": ["done"]
      },
      "DeleteResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      }
    }
  }
}